rockburst-app/requirements.txt

scikit-learn
xgboost
matplotlib
pillow
numpy
pandas
